Moving ProgressBarComponent to beta folder
