+++ app/components/primer/beta/progress_bar.html.erb
