+++ static/statuses.json
@@ -78,7 +78,8 @@
   "Primer::OcticonComponent": "beta",
   "Primer::OcticonSymbolsComponent": "alpha",
   "Primer::PopoverComponent": "beta",
-  "Primer::ProgressBarComponent": "beta",
+  "Primer::ProgressBarComponent": "deprecated",
+  "Primer::Beta::ProgressBar": "beta",
   "Primer::SpinnerComponent": "beta",
   "Primer::StateComponent": "beta",
   "Primer::SubheadComponent": "beta",
