+++ test/components/beta/progress_bar_test.rb
@@ -2,23 +2,23 @@
 
 require "components/test_helper"
 
-class Primer::ProgressBarComponentTest < Minitest::Test
+class PrimerBetaProgressBarTest < Minitest::Test
   include Primer::ComponentTestHelpers
 
   def test_does_not_render_if_no_items_provided
-    render_inline(Primer::ProgressBarComponent.new)
+    render_inline(Primer::Beta::ProgressBar.new)
 
     refute_component_rendered
   end
 
   def test_renders_empty_bar_if_percentage_is_not_provided
-    render_inline(Primer::ProgressBarComponent.new, &:item)
+    render_inline(Primer::Beta::ProgressBar.new, &:item)
 
     assert_selector("span.Progress .Progress-item")
   end
 
   def test_renders_large_option
-    render_inline(Primer::ProgressBarComponent.new(size: :large)) do |component|
+    render_inline(Primer::Beta::ProgressBar.new(size: :large)) do |component|
       component.item(percentage: 80)
     end
 
@@ -27,14 +27,14 @@ class Primer::ProgressBarComponentTest < Minitest::Test
 
   def test_renders_default_when_invalid_size_arg_passed
     without_fetch_or_fallback_raises do
-      render_inline(Primer::ProgressBarComponent.new(size: "kittens"), &:item)
+      render_inline(Primer::Beta::ProgressBar.new(size: "kittens"), &:item)
 
       assert_selector("span.Progress")
     end
   end
 
   def test_renders_percent_completed_progress
-    render_inline(Primer::ProgressBarComponent.new) do |component|
+    render_inline(Primer::Beta::ProgressBar.new) do |component|
       component.item(percentage: 80)
     end
 
@@ -42,7 +42,7 @@ class Primer::ProgressBarComponentTest < Minitest::Test
   end
 
   def test_renders_custom_styles
-    render_inline(Primer::ProgressBarComponent.new) do |component|
+    render_inline(Primer::Beta::ProgressBar.new) do |component|
       component.item(percentage: 80, style: "color: red")
     end
 
@@ -50,7 +50,7 @@ class Primer::ProgressBarComponentTest < Minitest::Test
   end
 
   def test_renders_background_colors
-    render_inline(Primer::ProgressBarComponent.new) do |component|
+    render_inline(Primer::Beta::ProgressBar.new) do |component|
       component.item(bg: :danger)
     end
 
@@ -58,6 +58,6 @@ class Primer::ProgressBarComponentTest < Minitest::Test
   end
 
   def test_status
-    assert_component_state(Primer::ProgressBarComponent, :beta)
+    assert_component_state(Primer::Beta::ProgressBar, :beta)
   end
 end
